+++ src/main/java/com/by/json/ShopJson.java
@@ -9,44 +9,44 @@ import java.util.stream.Collectors;
  * Created by yagamai on 15-12-9.
  */
 public class ShopJson {
-    private Long id;
-    private String name;
-    private String key;
-    private Integer[] menus;
+	private int id;
+	private String name;
+	private String key;
+	private Integer[] menus;
 
-    public ShopJson() {
-    }
+	public ShopJson() {
+	}
 
-    public ShopJson(Shop shop) {
-        this.id = shop.getId();
-        this.name = shop.getName();
-        this.key = shop.getShopKey();
-        this.menus = shop.getMenus().stream().map(Menu::getId).collect(Collectors.toList()).toArray(new Integer[]{});
-    }
+	public ShopJson(Shop shop) {
+		this.id = shop.getId();
+		this.name = shop.getName();
+		this.key = shop.getShopKey();
+		this.menus = shop.getMenus().stream().map(Menu::getId).collect(Collectors.toList()).toArray(new Integer[] {});
+	}
 
-    public Long getId() {
-        return id;
-    }
+	public int getId() {
+		return id;
+	}
 
-    public void setId(Long id) {
-        this.id = id;
-    }
+	public void setId(int id) {
+		this.id = id;
+	}
 
-    public String getName() {
-        return name;
-    }
+	public String getName() {
+		return name;
+	}
 
-    public void setName(String name) {
-        this.name = name;
-    }
+	public void setName(String name) {
+		this.name = name;
+	}
 
-    public String getKey() {
-        return key;
-    }
+	public String getKey() {
+		return key;
+	}
 
-    public void setKey(String key) {
-        this.key = key;
-    }
+	public void setKey(String key) {
+		this.key = key;
+	}
 
 	public Integer[] getMenus() {
 		return menus;
