+++ src/main/java/com/by/repository/ShopRepository.java
@@ -7,6 +7,6 @@ import com.by.model.Shop;
 /**
  * Created by yagamai on 15-12-1.
  */
-public interface ShopRepository extends PagingAndSortingRepository<Shop, Long> {
+public interface ShopRepository extends PagingAndSortingRepository<Shop, Integer> {
     Shop findByShopKey(String code);
 }
